refactor: moved s3-service into correct subfolder; updated references
